% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Correlation table}
\usage{
cor_test(
  data,
  cols,
  ...,
  digits = 3,
  quite = F,
  method = "pearson",
  p_adjust = "holm",
  return_result = F
)
}
\arguments{
\item{data}{data frame}

\item{cols}{correlation items. Support \code{dplyr::select()} syntax.}

\item{...}{additional arguments passed to correlation::correlation(). See ?correlation::correlation. Note that the return data.frame from correlation::correlation() must contains \code{r} and \code{p} (e.g., passing \code{baysesian = T} will not work)}

\item{digits}{number of digits to round to}

\item{quite}{suppress printing output}

\item{return_result}{If it is set to \code{TRUE}, it will return the data frame of the correlation table}
}
\value{
data frame of the correlation table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function uses the psych::corr.test function to generated the Pearson correlation table and their associated significance values.
}
\examples{
cor_test(iris, where(is.numeric))

}
