% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-check.R
\name{data_check}
\alias{data_check}
\title{Data Check (internal use only)}
\usage{
data_check(data)
}
\arguments{
\item{data}{data frame}
}
\description{
Data Check (internal use only)
}
\keyword{internal}
