% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_summary.R
\name{efa_summary}
\alias{efa_summary}
\title{Exploratory Factor Analysis}
\usage{
efa_summary(
  data,
  cols,
  rotation = "oblimin",
  n_factors = "default",
  scree_plot,
  fit_measures = T,
  factor_summary = T,
  bartlett_test = T,
  kmo_test = T
)
}
\arguments{
\item{data}{data frame}

\item{cols}{columns. dplyr::select syntax.}

\item{rotation}{the rotation to use in estimation. Default is oblimin. Options are 'none', 'varimax', 'quartimax', 'promax', 'oblimin', or 'simplimax'}

\item{n_factors}{the number of factors in the model. Default is data-driven estimation of most appropriate number of factors}

\item{scree_plot}{show explained variance by number of factor plot. default is \code{T}.}

\item{fit_measures}{show model fit measures and test. Default is \code{T}}

\item{factor_summary}{show factor summary showing the explained variance for each factor. Default is \code{T}}

\item{bartlett_test}{show Bartlett's test of sphericity result (you want this to be significant). Default is \code{T}}

\item{kmo_test}{show Kaiser-Meyer-Olkin (KMO) measure of sampling adequacy (MSA) results (you want the overall MSA to be higher than 0.7). Default is \code{T}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
The function fits a exploratory factor analysis model using the jmv::efa (The Jamovi Project, 2021) function. Users can fit the model by passing items using dplyr::select syntax. In the future, I may re-write this to be consisent with the layout of this package.
}
\details{
As a thumb of rule, you want factor loadings to be above 0.5.
}
\examples{
efa_summary(lavaan::HolzingerSwineford1939, starts_with("x"), scree_plot = TRUE)
}
\references{
The Jamovi Project. (2021). jmv: The 'jamovi' Analyses. R package version 1.2.23. Retrieved from https://github.com/jamovi/jmv.
}
