% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-format-round.R
\name{format_round}
\alias{format_round}
\title{Format digits (internal use only)}
\usage{
format_round(x, digits = 3)
}
\arguments{
\item{x}{object}

\item{digits}{number of digits to round to}
}
\value{
a rounded string
}
\description{
Format digits (internal use only)
}
\keyword{internal}
