% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-interaction-term.R
\name{get_interaction_term}
\alias{get_interaction_term}
\title{get interaction term}
\usage{
get_interaction_term(model)
}
\arguments{
\item{model}{model}
}
\value{
a list with predict vars names
}
\description{
get interaction term
}
