% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get-predict-df.R
\name{get_predict_df}
\alias{get_predict_df}
\title{get factor df to combine with mean_df}
\usage{
get_predict_df(data)
}
\arguments{
\item{data}{data}
}
\value{
factor_df
}
\description{
get factor df to combine with mean_df
}
