% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model_explore.R
\name{glm_model_explore}
\alias{glm_model_explore}
\title{Exploratory Linear Regression Model Table}
\usage{
glm_model_explore(
  data,
  response_variable,
  predictor_variable,
  family,
  two_way_interaction_variable = NULL,
  three_way_interaction_variable = NULL,
  control_variable = NULL,
  marginal_alpha = 0.1,
  verbose = TRUE,
  show_p = TRUE,
  return_result = FALSE,
  print_control = FALSE,
  plot_interaction = FALSE,
  file_dir = NULL,
  device = "jpeg",
  width = 8.5,
  height = 5,
  units = "in",
  y_lim = c(0, 1)
)
}
\arguments{
\item{data}{\code{data.frame}}

\item{response_variable}{Response variable. Support \code{dplyr::select()} syntax.}

\item{predictor_variable}{Pred variable. Support \code{dplyr::select()} syntax.}

\item{family}{a GLM family. It will passed to the family argument in glm See \code{?glm} for possible options.}

\item{two_way_interaction_variable}{Two-way interaction variable. Each two-way interaction variable will interact with each pred variable. Support \code{dplyr::select()} syntax.}

\item{three_way_interaction_variable}{Three-way interaction variable. Each three-way interaction variable will interact with each pred and two-way interaction variables. Support \code{dplyr::select()} syntax.}

\item{control_variable}{Control variables. Support \code{dplyr::select()} syntax.}

\item{marginal_alpha}{Set marginal_alpha level for marginally significant (denoted by \code{.}). Set to 0.05 if do not want marginally significant denotation.}

\item{verbose}{Default is \code{TRUE}. Set to \code{FALSE} to suppress outputs}

\item{show_p}{Default is \code{TRUE}. When \code{TRUE}, show the p-value in parenthesis.}

\item{return_result}{Default is \code{FALSE}. If \code{TRUE}, it returns the model estimates as a data frame.}

\item{print_control}{Default is \code{FALSE}. If \code{TRUE}, print coefficients of control variables.}

\item{plot_interaction}{Default is \code{FALSE}. If \code{TRUE}, interactions will be plotted and saved on your laptop.}

\item{file_dir}{Path of the directory to save plot to}

\item{device}{Default is "jpeg". See ggplot2::ggsave() for all options.}

\item{width}{Default is 8.5 (i.e., letter size width).}

\item{height}{Default is 5.}

\item{units}{Default is inches. Options are "in", "cm", "mm" or "px".}

\item{y_lim}{the plot's upper and lower limit for the y-axis. Length of 2. Example: \code{c(lower_limit, upper_limit)}}
}
\value{
data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Exploratory analyses for linear regression models with multiple response, predictor, and two-way interaction variables. (\code{lm} models).
At the moment, multi-categorical variables are not supported as predictors or interactions (but control is fine). Binary variable should be \code{numeric} instead of \code{factor}
}
\examples{

test = data.frame(y1 = sample(c(0, 1), size = 1000, replace = TRUE),
y2 = sample(c(0, 1), size = 1000, replace = TRUE),
y3 = sample(c(0, 1), size = 1000, replace = TRUE),
x1 = rnorm(1000,100,10),
x2 = rnorm(1000,10,1),
x3 = rnorm(1000,6,2),
m1 = rnorm(1000,3,1),
m2 = rnorm(1000,2,0.5),
m3 = rnorm(1000,9,0.1),
c1 = rnorm(1000,5,0.4),
c2 = rnorm(1000,2,0.2),
c3 = rnorm(1000,7,0.9)
)

glm_model_explore(data = test, 
                 response_variable = c(y1,y2,y3),
                 predictor_variable = c(x1,x2,x3),
                 two_way_interaction_variable = c(m1,m2,m3),
                 family = binomial(link = 'logit'),
                 control_variable = c(c1,c2,c3))
}
