% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme_model.R
\name{glme_model}
\alias{glme_model}
\title{Generalized Linear Mixed Effect Model}
\usage{
glme_model(
  data,
  model = NULL,
  response_variable,
  random_effect_factors = NULL,
  non_random_effect_factors = NULL,
  family,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  estimation_method = "REML",
  opt_control = "bobyqa",
  na.action = stats::na.exclude,
  quite = F
)
}
\arguments{
\item{data}{data frame}

\item{model}{lme4 model syntax. Support more complicated model. Note that model_summary will only return fixed effect estimates}

\item{response_variable}{character or vector of length 1}

\item{random_effect_factors}{vector. Level-1 variables (e.g., individual-level)}

\item{non_random_effect_factors}{optional vector. level-2 variables (e.g., group-level)}

\item{family}{a GLM family. It will passed to the family argument in glmer. See \code{?glmer} for possible options.}

\item{two_way_interaction_factor}{optional vector of length more than 2. Default to \code{null}}

\item{three_way_interaction_factor}{optional vector of length 3. Do not include two-way interaction factors if this is not null. Default to \code{null}}

\item{id}{character or vector of length 1. The nesting variable (e.g. group)}

\item{estimation_method}{character. \code{ML} or \code{REML} default to \code{REML}.}

\item{opt_control}{character. default to \code{bobyqa}}

\item{na.action}{default to \code{stats::na.exclude}.}

\item{quite}{suppress printing output}
}
\value{
An object of class "glmerMod" representing the linear mixed-effects model fit.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Fit a generalized linear mixed effect model using \code{lme4::glmer}. This function is still in early development stage.
}
\examples{
fit <- glme_model(
  response_variable = incidence,
  random_effect_factors = period,
  family = 'poisson', # or you can enter as poisson(link = 'log')
  id = herd,
  data = lme4::cbpp
)
}
