% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interaction-check.R
\name{interaction_check}
\alias{interaction_check}
\title{interaction_check}
\usage{
interaction_check(interaction_term)
}
\arguments{
\item{interaction_term}{}
}
\value{
nothing
}
\description{
interaction_check
}
\keyword{internal}
