% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interaction-value.R
\name{interaction_value}
\alias{interaction_value}
\title{Interaction_value}
\usage{
interaction_value(model_data, variable)
}
\arguments{
\item{model_data}{a data.frame}

\item{variable}{variable name}
}
\value{
For continuous variables, it returns a vector of -1SD and +1SD. For binary variables, it returns the two unique levels.
}
\description{
Utility function for extracting interaction values
}
\keyword{internal}
