% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_to_Rmd.R
\name{knit_to_Rmd}
\alias{knit_to_Rmd}
\title{Knit Rmd Files Instruction}
\usage{
knit_to_Rmd()
}
\value{
no return value
}
\description{
This is a helper function that instruct users of the package how to knit a R Markdown (Rmd) files
}
\examples{
knit_to_Rmd()
}
