% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label-name.R
\name{label_name}
\alias{label_name}
\title{get label name}
\usage{
label_name(
  graph_label_name,
  response_var_name,
  predict_var1_name,
  predict_var2_name,
  predict_var3_name
)
}
\arguments{
\item{graph_label_name}{label name}

\item{response_var_name}{outcome variable  name}

\item{predict_var1_name}{predictor 1 name}

\item{predict_var2_name}{predictor 2 name}

\item{predict_var3_name}{predictor 3 name}
}
\value{
vector of var name
}
\description{
get label name
}
