% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_model_explore.R
\name{lme_model_explore}
\alias{lme_model_explore}
\title{Exploratory Linear Mixed Effect Model Table}
\usage{
lme_model_explore(
  ...,
  data,
  response_variable,
  predictor_variable,
  two_way_interaction_variable = NULL,
  random_effect,
  control_variable = NULL,
  marginal_alpha = 0.1,
  return_result = FALSE,
  print_control = FALSE,
  verbose = TRUE,
  show_p = TRUE
)
}
\arguments{
\item{...}{additional parameters pass to lmerTest::lmer()}

\item{data}{\code{data.frame}}

\item{response_variable}{Response variable. Support \code{dplyr::select()} syntax.}

\item{predictor_variable}{Pred variable. Support \code{dplyr::select()} syntax.}

\item{two_way_interaction_variable}{Two-way interaction variable. Each two-way interaction variable will interact with each pred variable. Support \code{dplyr::select()} syntax.}

\item{random_effect}{The random-effects terms in the format of \verb{(|)}. See lm4::lmer for specifics.}

\item{control_variable}{Control variables. Support \code{dplyr::select()} syntax.}

\item{marginal_alpha}{Set marginal_alpha level for marginally significant (denoted by \code{.}). Set to 0.05 if do not want marginally significant denotation.}

\item{return_result}{Default is \code{FALSE}. If \code{TRUE}, it returns the model estimates as a data frame.}

\item{print_control}{Default is \code{FALSE}. If \code{TRUE}, print coefficients of control variables.}

\item{verbose}{Default is \code{TRUE}. Set to \code{FALSE} to suppress outputs}

\item{show_p}{Default is \code{TRUE}. When \code{TRUE}, show the p-value in parenthesis.}
}
\value{
data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Exploratory analyses for linear regression models with multiple response, predictor, and two-way interaction variables. (\code{lmer} models).
At the moment, multi-categorical variables are not supported as predictors or interactions (but control is fine). Binary variable should be \code{numeric} instead of \code{factor}
This function also do not supports changing random slopes.
}
\examples{

lme_model_explore(data = popular,
                  response_variable = c(popular,extrav),
                  predictor_variable = c(texp),
                  two_way_interaction_variable = sex,
                  random_effect = '(1 | class)')


}
