% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_model_table.R
\name{lme_model_table}
\alias{lme_model_table}
\title{Linear Mixed Effect Model Table}
\usage{
lme_model_table(
  ...,
  data,
  response_variable,
  predictor_variable,
  two_way_interaction_variable = NULL,
  random_effect,
  control_variable = NULL,
  other_parameters = NULL,
  marginal_alpha = 0.1,
  return_result = FALSE,
  verbose = TRUE,
  show_p = FALSE
)
}
\arguments{
\item{...}{additional parameters pass to lmerTest::lmer()}

\item{data}{\code{data.frame}}

\item{response_variable}{response variable. Support \code{dplyr::select()} syntax.}

\item{predictor_variable}{predictor variable. Support \code{dplyr::select()} syntax. It will automatically remove the response variable from predictor variable, so you can use \code{contains()} or \code{start_with()} safely.}

\item{two_way_interaction_variable}{Two-way interaction variable. Each two-way interaction variable will interact with the predictor variable. Support \code{dplyr::select()} syntax.}

\item{random_effect}{The random-effects terms in the format of \verb{(|)}. See lm4::lmer for specifics.}

\item{control_variable}{control variables. Support \code{dplyr::select()} syntax.}

\item{other_parameters}{catch call for all other parameters that need to be entered (e.g., non-changing interaction terms). Have to be \code{character} type.}

\item{marginal_alpha}{the set marginal_alpha level for marginally significant (denoted by \code{.}). Set to 0.05 if do not want marginally significant denotation.}

\item{return_result}{It set to \code{TRUE}, it return the model estimates data frame.}

\item{verbose}{default is \code{TRUE}. Set to \code{FALSE} to suppress outputs}

\item{show_p}{show the p-value in parenthesis}
}
\value{
data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Generate tables with multiple response, predictor, or two-way interaction variables (only \code{lmer} models are supported).
You can pass multiple variables for one type of variable (either response, pred, or interaction) only.
If you want to pass multiple variables for multiple type of variable, try lmer_model_explore instead.
At the moment, multi-categorical variables are not supported as predictors or interactions (but control is fine). Binary variable should be \code{numeric} instead of \code{factor}
This function also do not supports changing random slopes.
Please use \code{other_parameters} if you want to add non-changing interaction term.
}
\examples{
# If you want all varibles to be changing, try lmer_model_explore.
# For more examples, see ?lm_model_table. 

# Changing interaction terms with a non-changing response variable
lme_model_table(data = popular,
                 response_variable = popular,
                 predictor_variable = texp,
                 two_way_interaction_variable = c(extrav,sex),
                 random_effect = '(1 | class)')

# A non-changing interaction term with changing response variables
lme_model_table(data = popular,
                 response_variable = c(popular,sex),
                 predictor_variable = texp,
                 other_parameters = 'texp*extrav',
                 random_effect = '(1 | class)')
}
