% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_summary.R
\name{mediation_summary}
\alias{mediation_summary}
\title{Mediation Analysis}
\usage{
mediation_summary(
  data,
  response_variable,
  mediator,
  predictor_variable,
  control_variable = NULL,
  group = NULL,
  standardize = T,
  plot = F,
  digits = 3,
  return_result = F
)
}
\arguments{
\item{data}{data frame}

\item{response_variable}{response variable. Support \code{dplyr::select} syntax.}

\item{mediator}{mediator. Support \code{dplyr::select} syntax.}

\item{predictor_variable}{predictor variable. Support \code{dplyr::select} syntax.}

\item{control_variable}{control variables. Support \code{dplyr::select} syntax.}

\item{group}{nesting variable for multilevel mediation. Not confident about the implementation method. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}

\item{standardize}{standardized coefficients. Default is \code{TRUE}}

\item{plot}{plot using semPlot::semPath. I haven't found a nice-looking way to plot mediation using semPlot. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr}

\item{digits}{number of digits to round to}

\item{return_result}{Default is \code{FALSE}. If it is \code{TRUE}, it will return the \code{lavaan} object}
}
\value{
an object from \code{lavaan}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
It currently only support simple mediation analysis.
In the backend, it called the lavaan::sem() model. I am trying to implement multilevel mediation in \code{lavaan}.
In the future, I will try supporting moderated mediation (through \code{lavaan} or \code{mediation}) and mediation with latent variable (through \code{lavaan}).
There is a promising package called \code{brms} that use bayesian statistics. If you are interested, you can check that one out.
}
\examples{
mediation_summary(
  data = lmerTest::carrots,
  response_variable = Preference,
  mediator = Sweetness,
  predictor_variable = Crisp
)
}
