% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{model_summary}
\alias{model_summary}
\title{Model Summary for Mixed Effect Model}
\usage{
model_summary(
  model,
  streamlined_output = FALSE,
  round = 3,
  assumption_plot = FALSE,
  return_result = FALSE,
  quite = FALSE
)
}
\arguments{
\item{model}{an object from nlme::lme, lmerTest::lmer, or lme4::glmer}

\item{streamlined_output}{Only super_print model estimate and model performance. Default is \code{FALSE}}

\item{round}{number of digit to round the values.}

\item{assumption_plot}{Generate an panel of plots that check major assumptions. You can use this if the model summary show violation of assumption (those maybe unreliable due to the use of p-value which is sensitive to the sample size). In the background, it calls performance::check_model()}

\item{return_result}{return the model estimates data frame. Default is \code{FALSE}}

\item{quite}{suppress printing output}
}
\value{
If return_result is \code{TRUE}, it will return a data frame with estimate, df, p_value, and the p-value significance.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
The function will extract the relevant coefficients from the linear mixed effect models (see supported model below).
}
\details{
If you working with \code{lm}, fixed factor estimate is just the model estimate. It
}
\examples{
# I am going to show the more generic usage of this function
# You can also use this package's built in function to fit the models
# I recommend using the model_summary_with_plot to get everything

# lme example
lme_fit <- lme4::lmer("popular ~ texp  + (1 | class)",
  data = popular
)

model_summary(lme_fit) 

# lm example

lm_fit <- lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width,
  data = iris
)

model_summary(lm_fit, assumption_plot = TRUE)
}
\references{
Nakagawa, S., & Schielzeth, H. (2013). A general and simple method for obtaining R2 from generalized linear mixed-effects models. Methods in Ecology and Evolution, 4(2), 133–142. https://doi.org/10.1111/j.2041-210x.2012.00261.x
}
