% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary_with_plot.R
\name{model_summary_with_plot}
\alias{model_summary_with_plot}
\title{Model Summary with Interaction Plot}
\usage{
model_summary_with_plot(
  data,
  response_variable,
  level_1_factors,
  level_2_factors = NULL,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  cateogrical_var = NULL,
  id,
  family = NULL,
  graph_label_name = NULL,
  estimation_method = "REML",
  opt_control = "optim",
  na.action = na.exclude,
  model_performance = c("R2_fixed_effect", "R2_full_model"),
  return_result = NULL,
  print_result = c("short_summary", "plot"),
  y_lim = NULL,
  plot_color = F,
  estimate_round = 3,
  p_value_round = 3,
  performance_round = 3,
  use_package = "nlme",
  quite = F
)
}
\arguments{
\item{data}{required dataframe}

\item{response_variable}{required character or vector of length 1}

\item{level_1_factors}{vector. Level-1 variables (e.g., individual-level)}

\item{level_2_factors}{optional vector. level-2 variables (e.g., group-level)}

\item{two_way_interaction_factor}{optional vector of length more than 2. Default to `null`.}

\item{cateogrical_var}{list. use the form list(var_name1 = c(upper_bound1, lower_bound1), [var_name2 = c(upper_bound2, lower_bound2]).}

\item{id}{character or vector of length 1. The nesting variable (e.g. group).}

\item{family}{a GLM family. It will pased to the family argument in glmer. See `?glmer` for possible options. It will also call glmer.nb if family = 'negbin'}

\item{graph_label_name}{optional vector or function. vector of length 2 for two-way interaction graph. vector of length 3 for three-way interaction graph. Vector should be passed in the form of c(response_var, predict_var1, predict_var2, [predict_var3]). Function should be passed as a switch function. See below for an example.}

\item{estimation_method}{character. `ML` or `REML` default to `REML`.}

\item{opt_control}{character. default to `optim` for `lme` and `bobyqa` for lmerTest}

\item{na.action}{default to `na.exclude`.}

\item{model_performance}{vector. `R2_full_model` for conditional R2 and `R2_fixed_effect` for marginal R2 (Nakagawa, 2013). `icc` for intraclass correlation coefficient. The function calls the performance package for R2 and ICC (Lüdecke et al., 2020).}

\item{return_result}{optional vector. Choose from  `model`,`plot`,`short_summary`,`long_summary`,`bruceR_summary`.`model` return the model object. `plot` return the interaction plot.`short_summary` return a short model summary. `long_summary` return the summary. `bruceR_summary` uses the bruceR::HLM_summary (Bao, 2021) function.}

\item{print_result}{optional vector. Choose from `model`,`plot`,`short_summary`,`long_summary`.`model` return the model object. `plot` return the interaction plot.`short_summary` return a short model summary. `long_summary` return the summary.}

\item{y_lim}{vector of length 2. c(lower_limit, upper_limit)}

\item{plot_color}{logical. default as F. Set to T if you want to plot in color}

\item{estimate_round}{numeric. number of digit to round to for the slope estimate.}

\item{p_value_round}{numeric. number of digit to round to for the p-value.}

\item{performance_round}{numeric. number of digit to round to for the performance statistics (e.g., R2)}

\item{use_package}{character. Only avaliable for linear mixed effect model. Options are "nlme" or "lmerTest". Default is "nlme".}

\item{quite}{default to F. If set to `T`, it will not print the fitting model statement}

\item{three_way_interaction_factoroptional}{vector of length 3. Do not include two-way interaction factors if this is not null. Default to `null`.}
}
\value{
return a list of all requested items in the order of model, short_summary, long_summary, plot
}
\description{
It will first compute the mixed effect model. It will use either the nlme::lme ((Pinheiro, 2006) or the lmerTest::lmer (Kuznetsova, 2017) for linear mixed effect model. It will use lme4::glmer (Bates et al., 2014) for generalized linear mixed effect model. Then, it will graph the interaction using the two_way_interaction_plot or the three_way interaction_plot (Moy, 2021). If you requested simple slope summary, it will uses the interaction::sim_slopes (Long, 2019). If you requested bruceR_summary, it will uses the bruceR::HLM_summary (Bao, 2021)
}
\examples{

model_summary_with_plot(response_variable = 'incidence',
                        level_1_factors = 'size',
                        level_2_factors = 'herd',
                        two_way_interaction_factor = c('size','herd'),
                        id = 'period',
                        data = lme4::cbpp,
                        use_package = 'lmerTest', #use lmerTest
                        model_performance = NULL)

}
\references{
Bao, H.-W.-S. (2021). bruceR: Broadly useful convenient and efficient R functions. R package version 0.6.0. https://CRAN.R-project.org/package=bruceR

Bates, D., Mächler, M., Bolker, B., & Walker, S. (2014). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1–48. doi: 10.18637/jss.v067.i01.

Kuznetsova, A., Brockhoff, P. B., & Christensen, R. H. (2017). lmerTest package: tests in linear mixed effects models. Journal of statistical software, 82(13), 1-26.

Long J. A. (2019). interactions: Comprehensive, User-Friendly Toolkit for Probing Interactions. R package version 1.1.0, https://cran.r-project.org/package=interactions.

Lüdecke, D., Makowski, D., Waggoner, P., Patil I (2020). performance: Assessment of Regression Models Performance. CRAN. doi: 10.5281/zenodo.3952174, R package, https://easystats.github.io/performance/.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modelling in R. R package version 0.1.0, https://github.com/jasonmoy28/psycModel.#'

Nakagawa, S., & Schielzeth, H. (2013). A general and simple method for obtaining R2 from generalized linear mixed-effects models. Methods in Ecology and Evolution, 4(2), 133–142. https://doi.org/10.1111/j.2041-210x.2012.00261.x

Pinheiro, J., Bates, D., DebRoy, S., Sarkar, D., & Team, R. C. (2006). nlme: Linear and nonlinear mixed effects models. R package version, 3(4), 109.
}
