% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R
\name{model_table}
\alias{model_table}
\title{Model Table
Generate tables with multiple response and predictor variable (only \code{lm} models are supported)}
\usage{
model_table(
  data,
  response_variable,
  predictor_variable,
  control_variable = NULL,
  full_model = FALSE,
  alpha = 0.1,
  return_result = FALSE,
  quite = FALSE
)
}
\arguments{
\item{data}{\code{data.frame}}

\item{response_variable}{response variable. Support \code{dplyr::select()} syntax.}

\item{predictor_variable}{predictor variable. Support \code{dplyr::select()} syntax. It will automatically remove the response variable from predictor variable, so you can use \code{contains()} or \code{start_with()} safely.}

\item{control_variable}{control variables. Support \code{dplyr::select()} syntax.}

\item{full_model}{Default is \code{FALSE}. If \code{TRUE}, it will report the full model with coefficients of the control variables}

\item{alpha}{the set alpha level for marginally significant (denoted by \code{.}). Set to 0.05 if do not want marginally significant denotation.}

\item{return_result}{It set to \code{TRUE}, it return the model estimates data frame.}

\item{quite}{suppress printing output}
}
\value{
data.frame
}
\description{
Model Table
Generate tables with multiple response and predictor variable (only \code{lm} models are supported)
}
\examples{

model_table(data = iris, 
            response_variable = c(Sepal.Length,Sepal.Width),
            predictor_variable = Petal.Width)
}
