% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_regression_plot.R
\name{polynomial_regression_plot}
\alias{polynomial_regression_plot}
\title{Polynomial Regression Plot}
\usage{
polynomial_regression_plot(
  model,
  model_data = NULL,
  predictor,
  graph_label_name = NULL,
  x_lim = NULL,
  y_lim = NULL,
  plot_color = FALSE
)
}
\arguments{
\item{model}{object from \code{lm}}

\item{model_data}{optional dataframe (in case data cannot be retrieved from the model)}

\item{predictor}{predictor variable name (must be character)}

\item{graph_label_name}{vector of length 3 or function. Vector should be passed in the form of \code{c(response_var, predict_var1, predict_var2)}. Function should be passed as a switch function that return the label based on the name passed (e.g., a switch function)}

\item{x_lim}{the plot's upper and lower limit for the x-axis. Length of 2. Example: \code{c(lower_limit, upper_limit)}}

\item{y_lim}{the plot's upper and lower limit for the y-axis. Length of 2. Example: \code{c(lower_limit, upper_limit)}}

\item{plot_color}{default if \code{FALSE}. Set to \code{TRUE} if you want to plot in color}
}
\value{
an object of class \code{ggplot}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
The function create a simple regression plot (no interaction). Can be used to visualize polynomial regression.
}
\details{
It appears that \code{predict} cannot handle categorical factors. All variables are converted to numeric before plotting.
}
\examples{
fit = lm(data = iris, Sepal.Length ~ poly(Petal.Length,2))
polynomial_regression_plot(model = fit,predictor = 'Petal.Length')
}
