% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\name{Print}
\alias{Print}
\title{Print from bruceR (internal-use only)}
\usage{
Print(...)
}
\arguments{
\item{...}{...}
}
\description{
Print from bruceR (internal-use only)
}
\section{Functions}{
\itemize{
\item \code{Print}: Paste strings.
}}

