% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_summary.R
\name{reliability_summary}
\alias{reliability_summary}
\title{Reliability Analysis}
\usage{
reliability_summary(data, cols, digits = 3, CFA = T, descriptive_table = T)
}
\arguments{
\item{data}{data frame}

\item{cols}{items for reliability analysis.  Support \code{dplyr::select} syntax.}

\item{digits}{number of digits}

\item{CFA}{Run confirmatory factor analysis. Default is \code{TRUE}}

\item{descriptive_table}{Get descriptive statistics. Default is \code{TRUE}}
}
\description{
For multidimensional scale (default), it will uses the psych::omega function to get the omega total, omega hierarchical, and alpha. For uni-dimensional scale, it will compute the alpha and a single-factor CFA model.
}
\examples{
fit <- reliability_summary(
  data = lavaan::HolzingerSwineford1939,
  cols = x1:x3
)

}
