% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-super-print.R
\name{super_print}
\alias{super_print}
\title{super_print (internal use only)}
\usage{
super_print(text, env = parent.frame())
}
\arguments{
\item{text}{inputted text}

\item{env}{environment.}
}
\value{
formatted text (greek letter, color, font style formatted)
}
\description{
print color, greek-letter, font-style formatted text. Inputted text need to be 'command|$greek letter$'.
Command options are bold, italic, underline, red, green, yellow, blue, silver.
Greek letter options are alpha, beta, gamma, delta, epsilon, eta, theta, lamda, pi, rho, sigma, chi, omega. If capital letter greek letter, use all capital letter command (e.g., ALPHA)
If you wish to print normal text, then the inputted text should '|text'
}
\keyword{internal}
