% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-three-way-interaction-terms.R
\name{three_way_interaction_terms}
\alias{three_way_interaction_terms}
\title{Interaction term for Mixed Effect Model (internal use only)
Create interaction terms for regression models}
\usage{
three_way_interaction_terms(interaction_terms)
}
\arguments{
\item{interaction_terms}{vector. names of the interaction terms}
}
\value{
a vector of all three-way interaction terms
}
\description{
Interaction term for Mixed Effect Model (internal use only)
Create interaction terms for regression models
}
\keyword{internal}
