% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way_interaction_plot.R
\name{two_way_interaction_plot}
\alias{two_way_interaction_plot}
\title{Two-way Interaction Plot}
\usage{
two_way_interaction_plot(
  data,
  model,
  response_var,
  predict_var_name,
  graph_label_name = NULL,
  cateogrical_var = NULL,
  y_lim = NULL,
  plot_color = F
)
}
\arguments{
\item{data}{dataframe.}

\item{model}{lme object from `nlme::lme`}

\item{response_var}{response variable name}

\item{predict_var_name}{vector of length 2. predictive variable names}

\item{graph_label_name}{vector of length 3 or function. Vector should be passed in the form of c(response_var, predict_var1, predict_var2). Function should be passed as a switch function that return the label based on the name passed (e.g., a switch function)}

\item{cateogrical_var}{list. use the form list(var_name1 = c(upper_bound1, lower_bound1), [var_name2 = c(upper_bound2, lower_bound2])}

\item{y_lim}{vector of length 2. c(lower_limit, upper_limit)}

\item{plot_color}{logical. default as F. Set to T if you want to plot in color}
}
\value{
ggplot object. two-way interaction plot
}
\description{
The function creates a two-way interaction plot. It will creates a plot with ± 1 SD from the mean of the independent variable.
}
\examples{


}
\references{
Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modelling in R. R package version 0.1.0, https://github.com/jasonmoy28/psycModel.#'
}
