% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_terms.R
\name{two_way_interaction_terms}
\alias{two_way_interaction_terms}
\title{Interaction term for Mixed Effect Model (do not call directly)
Create Interaction terms (internal-use only)}
\usage{
two_way_interaction_terms(interaction_terms)
}
\arguments{
\item{interaction_terms}{vector. names of the interaction terms}
}
\value{
vector. The vector is essential for the lme formula used in function \code{HLM.Model::HLM_Model}
}
\description{
Interaction term for Mixed Effect Model (do not call directly)
Create Interaction terms (internal-use only)
}
\keyword{internal}
