% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_groupwise.R
\name{cfa_groupwise}
\alias{cfa_groupwise}
\title{Confirmatory Factor Analysis (groupwise)}
\usage{
cfa_groupwise(data, ..., model = NULL, group, ordered = F)
}
\arguments{
\item{data}{data frame}

\item{...}{CFA items.}

\item{model}{explicit lavaan model. Must be specify with \code{model = lavaan_model_syntax}.}

\item{group}{character. group variable.}

\item{ordered}{logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function will run a series of CFA (n = length(group)) with respect to each group. The function is intended to help you get a better understanding of which group has abnormal fit indicator
}
\details{
All argument must be explicitly specified. If not, all arguments will be treated as CFA items
}
\examples{
# The example is used as the illustration of the function output only.
# It does not imply the data is appropriate for the analysis.
cfa_groupwise(
  data = lavaan::HolzingerSwineford1939,
  group = "school",
  x1:x3,
  x4:x6,
  x7:x9
)
}
\references{
Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Rosseel Y (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1–36. https://www.jstatsoft.org/v48/i02/.
}
