% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coefficent-to-p.R
\name{coefficent_to_p}
\alias{coefficent_to_p}
\title{change coefficient to p value for model_table}
\usage{
coefficent_to_p(data_frame, marginal_alpha = 0.1, show_p = FALSE)
}
\arguments{
\item{data_frame}{data}

\item{marginal_alpha}{marginal_alpha leve}

\item{show_p}{show por not}
}
\description{
change coefficient to p value for model_table
}
\keyword{internal}
