% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fit.R
\name{compare_fit}
\alias{compare_fit}
\title{Comparison of Model Fit}
\usage{
compare_fit(models, digits = 3)
}
\arguments{
\item{models}{list of object from \code{lavaan} (for measurement invariance only at the moment) or lme, lmer models.}

\item{digits}{number of digit to round}
}
\value{
data frame with fit indices and change in fit indices
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Compare the fit indices models (see below for model support)
}
\examples{
# lme model
fit1 = lme_model(
  response_variable = JS_Individual,
  level_1_factors = Age_Individual,
  level_2_factors = Hofstede_IC_Country,
  id = Country,
  data = EWCS_2015_shorten,
)

fit2 = lme_model(
  response_variable = JS_Individual,
  level_1_factors = Age_Individual,
  level_2_factors = c(Hofstede_IC_Country,Hofstede_UA_Country),
  id = Country,
  data = EWCS_2015_shorten,
)

compare_fit(list(fit1, fit2))

# see ?measurement_invariance for measurement invariance example

}
