% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fit.R
\name{compare_fit}
\alias{compare_fit}
\title{Comparison of Model Fit}
\usage{
compare_fit(..., digits = 3)
}
\arguments{
\item{...}{model. If it is a lavaan object, it will try to perform measurement invariance. Other model type will passed to performance::compare_performance}

\item{digits}{number of digit to round}
}
\value{
data frame with fit indices and change in fit indices
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Compare the fit indices models (see below for model support)
}
\examples{
# lme model

fit1 <- lme_model(
  data = popular,
  response_variable = popular,
  random_effect_factors = extrav,
  non_random_effect_factors = texp,
  id = class
)

fit2 <- lme_model(
  data = popular,
  response_variable = popular,
  random_effect_factors = extrav,
  non_random_effect_factors = texp,
  two_way_interaction_factor = c(extrav, texp),
  id = class
)

compare_fit(fit1, fit2)

# see ?measurement_invariance for measurement invariance example
}
