% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fit.R
\name{compare_fit}
\alias{compare_fit}
\title{Comparison of Model Fit}
\usage{
compare_fit(models, digits = 3)
}
\arguments{
\item{models}{list of object from \code{lavaan} (potential expansion with other models)}

\item{digits}{number of digit to round}
}
\value{
data frame with fit indices and change in fit indices
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Compare the fit indices models (see below for model support)
}
