% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Correlation table}
\usage{
cor_test(
  data,
  cols,
  sig_test = "raw",
  digit = 3,
  quite = F,
  return_result = F,
  ...
)
}
\arguments{
\item{data}{data frame}

\item{cols}{columns. tidyselect syntax or helpers.}

\item{sig_test}{Default is raw. Options are 'adjusted' or 'raw'. Adjusted use holm adjustment method. See ?stats::p.adjust to learn why.}

\item{digit}{number of digits}

\item{quite}{suppress printing output}

\item{return_result}{return the data frame of the descriptive table}

\item{...}{additional argument for internal use}
}
\value{
data frame of the correlation table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function uses the psych::corr.test function to generated the Pearson correlation table and their associated significance values.
}
\examples{
cor_test(iris, where(is.numeric))
cor_test(iris, where(is.numeric), sig_test = "adjusted") # use adjusted correlation
}
