% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Correlation table}
\usage{
cor_test(data, cols, sig_test = "raw", digit = 3, ...)
}
\arguments{
\item{data}{a dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be recoded.}

\item{sig_test}{adjusted or raw. Default as adjusted. See psych::corr.test to learn more.}

\item{digit}{number of digits}
}
\description{
This function uses the psych::corr.test (Revelle, 2021) function to generated the pearson correlation table associated its significance values.
}
\examples{
cor_test(iris,1:4)
cor_test(iris,where(is.numeric))

}
\references{
Revelle, W. (2021). psych: Procedures for Psychological, Psychometric, and Personality Research. Northwestern University, Evanston, Illinois. R package version 2.1.3, https://CRAN.R-project.org/package=psych.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modelling in R. R package version 0.1.0, https://github.com/jasonmoy28/psycModel.
}
