% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Correlation table}
\usage{
cor_test(data, cols, sig_test = "raw", digit = 3, ...)
}
\arguments{
\item{data}{data frame}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be recoded.}

\item{sig_test}{Default is raw. Options are 'adjusted' or 'raw'. Adjusted use holm adjustment method. See ?stats::p.adjust to learn why.}

\item{digit}{number of digits}

\item{...}{additional argument.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function uses the psych::corr.test (Revelle, 2021) function to generated the pearson correlation table and their associated significance values.
}
\examples{
cor_test(iris, where(is.numeric))
cor_test(iris, where(is.numeric),sig_test = 'adjusted') # use adjusted correlation 

}
\references{
Revelle, W. (2021). psych: Procedures for Psychological, Psychometric, and Personality Research. Northwestern University, Evanston, Illinois. R package version 2.1.3, https://CRAN.R-project.org/package=psych.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel
}
