% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_table.R
\name{descriptive_table}
\alias{descriptive_table}
\title{Mean, SD, and Correlation table}
\usage{
descriptive_table(
  data,
  cols,
  cor_sig_test = "raw",
  cor_digit = 3,
  descriptive_indicator = c("mean", "sd", "cor"),
  descriptive_indicator_digit = 3,
  file_path = NULL
)
}
\arguments{
\item{data}{data frame}

\item{cols}{vector or tidyselect syntax or helpers. column(s) need to be included in the table.}

\item{cor_sig_test}{adjusted or raw. Default as adjusted. See psych::corr.test to learn more.}

\item{cor_digit}{number of digit for correlation table}

\item{descriptive_indicator}{Default is mean, sd, cor. Options are missing (missing value count), non_missing (non-missing value count), cor (correlation table), n, mean, sd, median, trimmed (trimmed mean), median, mad (median absolute deviation from the median), min, max, range, skew, kurtosis, se (standard error)}

\item{descriptive_indicator_digit}{number of digit for the descriptive table}

\item{file_path}{file path for export. The function will implicitly pass this argument to the write.csv(file = file_path)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function generates a table of descriptive statistics (mainly using psych::describe; Revelle, 2021) and or a correlation table. User can export this to a csv file (optionally, using the file_path argument). Users can open the csv file with MS Excel then copy and paste the table into MS Word table.
}
\examples{
descriptive_table(iris, cols = where(is.numeric)) # all numeric columns

descriptive_table(iris,
  cols = where(is.numeric),
  # get missing count, non-missing count, and mean & sd & correlation table
  descriptive_indicator = c("missing", "non_missing", "mean", "sd", "cor")
)
}
\references{
Revelle, W. (2021). psych: Procedures for Psychological, Psychometric, and Personality Research. Northwestern University, Evanston, Illinois. R package version 2.1.3, https://CRAN.R-project.org/package=psych.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel
}
