% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glme_model.R
\name{glme_model}
\alias{glme_model}
\title{Generalized Linear Mixed Effect Model}
\usage{
glme_model(
  data,
  response_variable,
  level_1_factors,
  level_2_factors = NULL,
  family,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  estimation_method = "REML",
  opt_control = "bobyqa",
  na.action = stats::na.exclude,
  quite = F
)
}
\arguments{
\item{data}{data frame}

\item{response_variable}{character or vector of length 1}

\item{level_1_factors}{vector. Level-1 variables (e.g., individual-level)}

\item{level_2_factors}{optional vector. level-2 variables (e.g., group-level)}

\item{family}{a GLM family. It will passed to the family argument in glmer. See \code{?glmer} for possible options.}

\item{two_way_interaction_factor}{optional vector of length more than 2. Default to \code{null}}

\item{three_way_interaction_factor}{optional vector of length 3. Do not include two-way interaction factors if this is not null. Default to \code{null}}

\item{id}{character or vector of length 1. The nesting variable (e.g. group)}

\item{estimation_method}{character. \code{ML} or \code{REML} default to \code{REML}.}

\item{opt_control}{character. default to \code{bobyqa}}

\item{na.action}{default to \code{stats::na.exclude}.}

\item{quite}{default to F. If set to \code{T}, it will not print the fitting model statement}
}
\value{
An object of class "glmerMod" representing the linear mixed-effects model fit.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Generalized linear mixed effect model. The function uses the \code{lme4::glmer}(Bates et al., 2014) function
}
\examples{
x <- rnorm(200)
y <- rpois(200, exp(1+x))
group = rep(1:10,20)
test_df = as.data.frame(cbind(x,y,group))
glme_model(data = test_df,
           response_variable = 'y',
           level_1_factors = 'x',
           family = poisson(link = 'log'),
           id = 'group')

}
\references{
Bates, D., Mächler, M., Bolker, B., & Walker, S. (2014). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1–48. doi: 10.18637/jss.v067.i01.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel
}
