% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_model.R
\name{lme_model}
\alias{lme_model}
\title{Linear Mixed Effect Model}
\usage{
lme_model(
  data,
  response_variable,
  level_1_factors,
  level_2_factors = NULL,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  estimation_method = "REML",
  opt_control = "optim",
  na.action = na.exclude,
  use_package = "nlme",
  quite = F
)
}
\arguments{
\item{data}{dataframe}

\item{response_variable}{character or vector of length 1}

\item{level_1_factors}{vector. Level-1 variables (e.g., individual-level)}

\item{level_2_factors}{optional vector. level-2 variables (e.g., group-level)}

\item{two_way_interaction_factor}{optional vector of length more than 2. Default to `null`}

\item{three_way_interaction_factor}{optional vector of length 3. Do not include two-way interaction factors if this is not null. Default to `null`}

\item{id}{character or vector of length 1. The nesting variable (e.g. group)}

\item{estimation_method}{character. `ML` or `REML` default to `REML`.}

\item{opt_control}{character. default to `optim` for `lme` and `bobyqa` for lmerTest}

\item{na.action}{default to `na.exclude`.}

\item{use_package}{character. `nlme` or `lmerTest`. Default is `nlme`.}

\item{quite}{default to F. If set to `T`, it will not print the fitting model statement}
}
\value{
An object of class `lme` of `lmerModLmerTest` representing the linear mixed-effects model fit.
}
\description{
Compute a linear mixed effect model (i.e., hierarchical linear model, multilevel linear model) using the `nlme::lme` (Pinheiro, 2006)  or the `lmerTest::lmer` (Kuznetsova, 2017) fucntion.
}
\examples{
lme_model(response_variable = 'Reaction',
           level_1_factors = 'Days',
           id = 'Subject',
           data = lme4::sleepstudy)


}
\references{
Kuznetsova, A., Brockhoff, P. B., & Christensen, R. H. (2017). lmerTest package: tests in linear mixed effects models. Journal of statistical software, 82(13), 1-26.

Moy, J. H. (2021). psycModel: Intergrated Toolkit for Pyschological Analysis and Modelling in R. Retrieved from https://github.com/jasonmoy28/psycModel.

Pinheiro, J., Bates, D., DebRoy, S., Sarkar, D., & Team, R. C. (2006). nlme: Linear and nonlinear mixed effects models. R package version, 3(4), 109.
}
