% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_model.R
\name{lme_model}
\alias{lme_model}
\title{Linear Mixed Effect Model}
\usage{
lme_model(
  data,
  response_variable,
  level_1_factors,
  level_2_factors = NULL,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  estimation_method = "REML",
  opt_control = "optim",
  na.action = stats::na.exclude,
  use_package = "nlme",
  quite = F
)
}
\arguments{
\item{data}{data frame}

\item{response_variable}{character or vector of length 1}

\item{level_1_factors}{vector. Level-1 variables (e.g., individual-level)}

\item{level_2_factors}{optional vector. level-2 variables (e.g., group-level)}

\item{two_way_interaction_factor}{optional vector of length more than 2. Default to \code{null}}

\item{three_way_interaction_factor}{optional vector of length 3. Do not include two-way interaction factors if this is not null. Default to \code{null}}

\item{id}{character or vector of length 1. The nesting variable (e.g. group)}

\item{estimation_method}{character. \code{ML} or \code{REML} default to \code{REML}.}

\item{opt_control}{character. default to \code{optim} for \code{lme} and \code{bobyqa} for lmerTest}

\item{na.action}{default to \code{stats::na.exclude}.}

\item{use_package}{character. \code{nlme} or \code{lmerTest}. Default is \code{nlme}.}

\item{quite}{default to F. If set to \code{T}, it will not print the fitting model statement}
}
\value{
An object of class \code{lme} of \code{lmerModLmerTest} representing the linear mixed-effects model fit.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Compute a linear mixed effect model (i.e., hierarchical linear model, multilevel linear model) using the \code{nlme::lme} (Pinheiro, 2006)  or the \code{lmerTest::lmer} (Kuznetsova, 2017) function.
}
\examples{
lme_model(response_variable = 'Reaction',
           level_1_factors = 'Days',
           id = 'Subject',
           data = lme4::sleepstudy)


}
\references{
Kuznetsova, A., Brockhoff, P. B., & Christensen, R. H. (2017). lmerTest package: tests in linear mixed effects models. Journal of statistical software, 82(13), 1-26.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Pinheiro, J., Bates, D., DebRoy, S., Sarkar, D., & Team, R. C. (2006). nlme: Linear and nonlinear mixed effects models. R package version, 3(4), 109.
}
