% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{model_summary}
\alias{model_summary}
\title{Model Summary for Mixed Effect Model}
\usage{
model_summary(
  model,
  model_performance = c("R2_fixed_effect", "R2_full_model"),
  estimate_round = 3,
  p_value_round = 3,
  performance_round = 3,
  model_fit = "none"
)
}
\arguments{
\item{model}{an object from nlme::lme, lmerTest::lmer, or lme4::glmer}

\item{model_performance}{vector. `R2_full_model` for conditional R2 and `R2_fixed_effect` for marginal R2 (Nakagawa, 2013). `icc` for intraclass correlation coefficient. The function calls the performance package for R2 and ICC (Lüdecke et al., 2020).}

\item{estimate_round}{numeric. number of digit to round to for the slope estimate}

\item{p_value_round}{numeric. number of digit to round to for the p-value}

\item{performance_round}{numeric. number of digit to round to for the performance statistics (e.g., R2)}

\item{model_fit}{if you know your model produce same result as one of the object specify in the model argument, then you can try pass it to the function. Do it at your own risk.}
}
\value{
a dataframe with estimate, degree of freedom, p_value, and whether that p-value is significant
}
\description{
The function will extract the relevant coefficients from the linear mixed effect models (see supported model below).
}
\examples{
fit = lme_model(response_variable = 'Reaction',
               level_1_factors = 'Days',
               id = 'Subject',
               data = lme4::sleepstudy)

model_summary(fit)


}
\references{
Lüdecke, D., Makowski, D., Waggoner, P., Patil I (2020). performance: Assessment of Regression Models Performance. CRAN. doi: 10.5281/zenodo.3952174, R package, https://easystats.github.io/performance/.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modelling in R. R package version 0.1.0, https://github.com/jasonmoy28/psycModel.#'

Nakagawa, S., & Schielzeth, H. (2013). A general and simple method for obtaining R2 from generalized linear mixed-effects models. Methods in Ecology and Evolution, 4(2), 133–142. https://doi.org/10.1111/j.2041-210x.2012.00261.x
}
