% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-print-table.R
\name{print_table}
\alias{print_table}
\title{print_table (internal use only)}
\usage{
print_table(data_frame, marginal_alpha = 0.1, digits = 3)
}
\arguments{
\item{data_frame}{data frame object}

\item{marginal_alpha}{marginal significant threshold}

\item{digits}{number of digits to round to}
}
\value{
no return value
}
\description{
Rich-text fitted table printing in console
}
\keyword{internal}
