% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way_interaction_plot.R
\name{three_way_interaction_plot}
\alias{three_way_interaction_plot}
\title{Three-way Interaction Plot}
\usage{
three_way_interaction_plot(
  data,
  model,
  response_var,
  predict_var_name,
  cateogrical_var = NULL,
  graph_label_name = NULL,
  y_lim = NULL,
  plot_color = F
)
}
\arguments{
\item{data}{dataframe}

\item{model}{lme object or lmerModLmerTest object}

\item{response_var}{response variable name}

\item{predict_var_name}{vector of length 3. the variables' name for the two-way interaction plot}

\item{cateogrical_var}{list. use the form list(var_name1 = c(upper_bound1, lower_bound1), [var_name2 = c(upper_bound2, lower_bound2])}

\item{graph_label_name}{vector of length 3 or function. Vector should be passed in the form of c(response_var, predict_var1, predict_var2, predict_var3). Function should be passed as a switch function that return the label based on the name passed (e.g., a switch function)}

\item{y_lim}{vector of two number. set the y_lim of the plot}

\item{plot_color}{logical. default as F. Set to T if you want to plot in color}
}
\value{
ggplot object. three-way interaction plot
}
\description{
The function creates a two-way interaction plot. It will creates a plot with ± 1 SD from the mean of the independent variable.
}
\examples{

}
\references{
Moy, J. H. (2021). psycModel: Intergrated Toolkit for Pyschological Analysis and Modelling in R. Retrieved from https://github.com/jasonmoy28/psycModel.
}
