% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_way_interaction_plot.R
\name{three_way_interaction_plot}
\alias{three_way_interaction_plot}
\title{Three-way Interaction Plot}
\usage{
three_way_interaction_plot(
  model,
  cateogrical_var = NULL,
  graph_label_name = NULL,
  y_lim = NULL,
  plot_color = F
)
}
\arguments{
\item{model}{lme, lmerMod, lmerModLmerTest object.}

\item{cateogrical_var}{list. Specify the upper bound and lower bound directly instead of using ± 1 SD from the mean. Passed in the form of \code{list(var_name1 = c(upper_bound1, lower_bound1),var_name2 = c(upper_bound2, lower_bound2))}}

\item{graph_label_name}{vector of length 3 or a switch function (see ?two_way_interaction_plot example). Vector should be passed in the form of c(response_var, predict_var1, predict_var2, predict_var3).}

\item{y_lim}{vector of two number. set the y_lim of the plot}

\item{plot_color}{logical. default as F. Set to T if you want to plot in color}
}
\value{
ggplot object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
The function creates a two-way interaction plot. It will creates a plot with ± 1 SD from the mean of the independent variable. See below for supported model
}
\examples{
fit = lme_model(response_variable = JS_Individual,
           level_1_factors = c(Age_Individual,Education_Individual), 
           level_2_factors = contains('Country'),
           three_way_interaction_factor = c('Age_Individual',
                                            'Education_Individual',
                                            'Hofstede_IC_Country'),
           id = Country,
           data = EWCS_2015_shorten)
           
three_way_interaction_plot(fit)           
           

}
\references{
Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel
}
